requirements.txt
asgiref==3.8.1
Django==5.0.4
lxml==5.2.1
numpy==1.26.4
pandas==2.2.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
XlsxWriter==3.2.0
